Use #to_f for better resolution

+++ lib/harness/counter.rb
@@ -25,7 +25,7 @@ module Harness
         counter.value = Harness.redis.incr(counter.id).to_i
       end
 
-      Harness.redis.zadd "meters/#{counter.id}", counter.time.to_i, counter.value
+      Harness.redis.zadd "meters/#{counter.id}", counter.time.to_f, counter.value
 
       counter
     end
